Tests scaffolding
Scaffold a test project with a web application test factory - FundaAssingment.Tests project. Configure HttpClient for stubbing. Create a new appsetting file for tests and have a setup that would use this new appsettings. Add env variable for API-KEY with "TEST" value.
Create a funda api json response builder for tests that will allow to shape the response from stub. It needs to give a possibility to configure a makelaar name per listing object and current page for paging prop