New TestServer per test run
I don't like that you re-register FundaApiClient in TestWebApplicationFactory, this way tests do not assess an original setup. Suggest a way to locate an existing IHttpClientBuilder in IOC to only attach a primary handler. Also, I want to recreate a WebAppTestFactory per test run with a new Http Mock behavior. Build expose a method that accepts a new Http Mock and calls WithWebHostBuilder under the hood.